Favorite symbol selection
This commit adds the possibility to select various symbols that will
be shown if the user holds down the Ctrl key while clicking the
right mouse button.
A pie/wheel menu will pop up (similar to the one of the frequently
used tools on just clicking the right mouse button) that allows to
select one of the symbols.