TagSelectWidget: Change column width
The 'Find objects' dialog contains a table which is shown in the
'Query editor' view.
The 'Relation' and 'Comparision' columns were set up with a fixed
width based on the width of the translated labels 'Relation'
and 'Comparision'. The result looks good for English but not
for other languages, e.g., German or Czech.
Set up both columns to automatically resize based on its contents.